Bug/older f string (#157)
CodeBuild, where our CDK code is run, is not actually using the latest
python version (3.12) on its latest platform version. I discovered this
when our CDK app stopped synthesizing this morning when it failed to
recognize a python 3.12 f-string feature use, introduced in the
monitoring-and-alerting feature. This PR resolves that issue.

### Description List
- ~Remove use of python 3.12 f-string feature use from CDK code~
*specify python version in the build environment*
- Add `python --version` call to ShellStep to explicitly print out
python version in use
- Remove main/development PR branch filter for python checks github
action (not strictly related to bug-fix but facilitating more consistent
check executions)