Add SES configuration with Cognito
In order for the cognito user pool to send emails
from the custom domain, we need to set up an SES
identity along with necessary DNS records to verify
that the domain may be used to send emails.
This also sets up an SNS topic which will forward
all bounce or complaint notifications to an
operations team email address.