Set context variable for ses operation email
The SES identity needs to be able to forward bounce
and complaint notifications for the domain, so that
an operations team can be made aware if a large spike
of message failures occurs for whatever reason. This
adds a required 'ses_feedback_forwarding_email' context
variable, which should be set to the email address of the
operations team that will manage these SES notifications.