Feat/support custom domain for cognito user email notifications (#198)
### Requirements List
- NOTE: This update requires an email address to receive SES email
bounce and complaint notifications. It must be added to the
cdk.context.json for each deployment environment under the following
field:

```
"environments": {
    "<env_name>": {
      "notifications": {
        "ses_operations_support_email": "<operation team email address here>"
....
```

### Description List
- In order for the cognito user pool to send emails
from the custom domain, we need to set up an SES
identity along with necessary DNS records to verify
that the domain may be used to send emails.
This also sets up an SNS topic which will forward
all bounce or complaint notifications to an
operations team email address.
- The SES identity needs to be able to forward bounce
and complaint notifications for the domain, so that
an operations team can be made aware if a large spike
of message failures occurs for whatever reason. This
adds a required 'ses_feedback_forwarding_email' context
variable, which should be set to the email address of the
operations team that will manage these SES notifications.
- The custom domain name is an optional field. If it is not provided,
the default cognito email settings will be used.

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- Code review

Closes #149