Move v1 api model definitions into api model class
To keep our schema setup separated from other configuration,
this moves all of the API GW model definitions for the v1 api
into their own class. This improves readability of the cdk logic
and sets up common source for multiple endpoints to share the same
model as needed.

This model is responsible for initializing the v1 api models
for the root api, which we then want to set as protected
so other classes won't modify it. This is a valid use case
for protected access to work with cdk.