Feat/licensee login (#216)
### Requirements List
- Use / deploy the API from
https://github.com/csg-org/CompactConnect/pull/208
- Add the following env vars to your .env file: 
```
      VUE_APP_COGNITO_AUTH_DOMAIN_LICENSEE=<your-licensee-auth-domain>
      VUE_APP_COGNITO_CLIENT_ID_LICENSEE=<your-licensee-userpool-client-id>
```
- Create a licensee user in your licensee user pool
- Create or verify that you have a staff user in your staff user pool

### Description List
- Modified /Login to have disambiguation links allowing users to choose
which user pool they should log in to rather than auto-redirecting
- Modified AuthCall page to differentiate between different user pools
and handle appropriately whether differentiating state param was
included or not
- Modified log in related user actions to function with arbitrary input
authType
- Added authType getter
- Added authType enum
- Modified loggedInGetter to reflect changes
- Modified Lambda that sets CSP settings for Cloudfront to include new
cognito domain
- Modified tests to reflect changes
- Modified dev workflow files to include env var

### Testing List
- Try logging in as licensee (can give you credentials on request) and
confirm that you are brought to the licenseedashboard with the
appropriate nav items
- View the tokens in session storage to confirm they are populated
appropriately
- Try logging in as staff and confirm that you are brought to the
appropriate page with the appropriate nav items
- View the tokens in session storage to confirm they are populated
appropriately
- Repeat the above 4 steps but remove the state query param from the url
bar once you are redirected to the cognito hosted login screen and
re-navigate and sign in, confirm the flows still work identically as
before
- login as staff then go directly to cognito login route for licensee
and login. You should be interacting with the app as though you are
staff. If you logout you should be brought to the staff login page. If
you then navigate directly back to the app you should still be logged in
as a licensee
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- Code review

### Notes:

- I did not add route "permissions" to this PR as the scope was
ballooning a bit and I wanted to get this chunk done. I believe John is
working on that in parallel
- In order to deliver this vertical chunk we have left the non-ideal
double login / logout flow to default in logging out of the highest
permissions account (staff). In the future we will use cognito's token
revocation endpoint to handle this edge case in whatever way
stakeholders deem most fit.
- This along with the changes in
https://github.com/csg-org/CompactConnect/pull/208 are together a
feature to add licensee login. I broke this out into its own PR because
it seemed like that PR was already underway and we will need to add the
test env vars that are only available after those changes are deployed.
- @jlkravitz I am putting this PR into draft until we have completed the
checklist below as that is necessary before merging. However, the rest
of the PR can be reviewed.

### Check List to move PR out of draft / needed before merge:
- [x] Merge https://github.com/csg-org/CompactConnect/pull/208 (Josh)
- [x] Get env vars from that environment, merge dev branch back into
this one, update appropriate github workflow file with env vars (Dana)
- [x] Get PR approval from John internally for the CSP changes fixes
(Dana)


Closes #201

---------

Co-authored-by: Dana Stiefel <dana@inspiringapps.com>