GET endpoint for Compact/Jurisdiction Privilege Fees (#243)
As part of this ticket:
https://github.com/csg-org/CompactConnect/issues/209 we have a need
for compacts and jurisdictions to define how much they plan to charge
for a provider
to purchase privileges. To save costs on our initial implementation, we
will **not** collect this information
through an API endpoint. Instead that information will need to be stored
within YAML configuration
files within this project. This commit adds documentation outlining the
onboarding process,
and provides configuration file placeholders

As part of this change, we are introducing a new DynamoDB table to hold
compact configuration.
This is because we want to keep our provider information separated from
our compact configuration
in terms of where it is stored. In order to upload the YAML
configuration in the Dynamo table, we have
added a CloudFormation custom resource, which is essentially a lambda
function that runs at deploy time
to upload this configuration.

Finally, this adds a new GET `/v1/purchases/privileges/options/`
endpoint, which will return all the jurisdiction and
compact fee information related to privileges. This will be used by the
frontend to display the list of available
jurisdiction privileges for a user to select from.

The high-level design for the YAML compact configuration files and
onboarding process was reviewed and
approved by @isabeleliassen during a previous coordination meeting. 


### Requirements List
- No new requirements needed to run deployment

### Description List
- Added YAML compact configuration files and onboarding documentation
- Added custom resource to convert YAML to JSON and upload configuration
into DynamoDB
- Updated CDK tests to leverage snapshots for large template
comparisons.
- Added GET `/v1/purchases/privileges/options/` endpoint in new
`purchases` directory.

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #209 
(backend work only, frontend will be done in separate PR)