Replacing pylint with ruff for checking python code
We have determined to add Ruff as our formatter for this project. Ruff can also be
used as a linter, and so we are replacing pylint to reduce the number of
dependencies we have to work with to maintain this project. This introduces the
Ruff dependency, and replaces the pylint commands with the Ruff check command.
See [Ruff tutorial][0]

This commit also includes several lint fixes caught by Ruff which were not
identified by pylint. We will need to clean up pylint comments throughout the
project as we update code in files.

[0]: https://docs.astral.sh/ruff/linter/