Feat/ruff formatter (#251)
We have determined to add Ruff as our formatter for this project. Ruff
can also be used as a linter, and so we are replacing pylint to reduce
the number of dependencies we have to work with to maintain this
project. This introduces the Ruff dependency, and replaces the pylint
commands with the Ruff check command. See [Ruff tutorial][0]

We have selected a reasonable set of linting/formatting rules to start
with, though we expect to adjust the rules of this tool as we continue
to use it. This PR introduces all of the formatting/linting recommended
changes so that every python file in the project is using a consistent
formatting according to this beginning rule set. Because this PR touches
every python file in the project, in the interest of avoiding
significant merge conflicts, it is requested that this PR not be held up
over decisions of specific ruff rules to apply, as those can be added in
subsequent PRs. The intent of this PR is to introduce the tooling with
consistent formatting applied.

**It is important to note that this PR is intended to be a pure refactor
with NO runtime code implementation changes.**

[0]: https://docs.astral.sh/ruff/linter/

### Requirements List
- Added new Ruff dependency to `compact-connect/requirements-dev.txt`,
which can be installed locally to run the
linter/formatter

### Description List
- Replace pylintrc with ruff.toml for linter/formatter configuration
rules.
- Ran automatic linter fixes, which includes the following:
   - sorting import in alphabetical order
   - removing redundant 'r' arg when opening files
- replacing deprecated type imports with recommended collections module
- Ran ruff formatter using recommended default options
- Replaced pylint comments with ruff comments where applicable

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #203

---------

Co-authored-by: Justin Frahm <justin@inspiringapps.com>