Feat/ Add endpoint for purchasing privileges (#272)
This PR includes a new endpoint `POST /v1/purchases/privileges` which
takes in a list of jurisdictions that a provider has
selected to purchase, as well as card billing information, to process a
transaction using the compact's payment processor
of choice. 

After the transaction is successfully completed, this endpoint then adds
the associated privilege records under the provider's list of
privileges.

We also have another ticket for adding an endpoint to store these
authorize.net credentials securely in secrets manager. See
https://github.com/csg-org/CompactConnect/issues/267

### Requirements List
- Until we complete
https://github.com/csg-org/CompactConnect/issues/267, these
authorize.net credentials will need to be stored manually in secrets
manager in order to test this endpoint in a specific AWS account.

### Description List
- Added `POST /v1/purchases/privileges` with lambda handler and tests

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #261