Extract initial common code into common directory (#300)
As part of #271 , this ticket addresses a technical need of sharing
common code across all of our lambda functions. We have opted to use a
lambda layer for this. As part of this change, all of the common schema,
client, and config files have been extracted into a `common-python`
directory, which we use to create the lambda layer in the persistent
stack. For local development/testing, we reference this common directory
in our unit tests by updating the PYTHONPATH env var to reference it,
which matches the behavior we should expect when the lambda function is
deployed with the layer attached to it.

It's important to note that CDK doesn't support including layers when
running hook tests, so we have determined to remove the test hook from
the bundle step. We currently run the tests as part of the PR review
phase, which must
pass in order for the code to be merged, so we still ensure passing unit
test coverage for all of our components.

### Requirements List
- no new requirements for deploying these changes

### Description List
- Extracted common python lambda code into `common-python` directory 
- Added lambda layer resource with reference to it in every v1 api
python lambda function

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #271

---------

Co-authored-by: Justin Frahm <justin@inspiringapps.com>