Calculating status for license during read
We are splitting the concept of license status into two terms.
The first is the status that is uploaded by the jurisdiction,
which we are tracking as the 'jurisdictionStatus'. The second
is the 'status' that is actually used by the system to determine
eligibility for a user to purchase privileges. This adds the
new 'jurisdictionStatus' field