Add dummy exports
CDK has a limitation where if you attempt to remove all references to
a resource from another stack, CDK will attempt to remove the export
for that resource, which then fails because it sees there is still an
external dependency on that resource, creating a 'deadly embrace'
See https://github.com/aws/aws-cdk/pull/12778.

The workaround for this is to set dummy export values to trick CDK
into keeping the unused exports around until you've deployed your
updates to the dependent stack.