Feat/Calculate license and privilege status based on expiration (#343)
The system needs to be aware of when a license or privilege is inactive.
Rather than depend on scheduled jobs to update the status of these
records, the status will be calculated whenever these records are loaded
from the DB.

We still need to track the status of the license as determined by the
jurisdiction which uploaded the license data into the system. This
introduces a new required field for licenses 'jurisdictionStatus' which
we will track in the system.

There is a chance that this status may not be updated by a jurisdiction
before it expires according to the 'dateOfExpiration' field on the
license in our system, so whenever we load the records for a license or
privilege, we calculate the status that is returned to the client by
checking the expiration against the current moment (we have determined
to use the US Eastern Time zone when determining the date. Though
management expects this to likely change in the future)

### Requirements List
- This introduces a new required field for licenses
'jurisdictionStatus', which is 🚨🚨 **a backwards incompatible change** 🚨🚨
with the
current records in the database. Given that we are currently only using
test records, we can clear the DB and load new test data.

### Description List
- Add schema loading hooks to calculate the status of licenses and
privileges whenever we read them from the DB.
- Move common testing resources into the common folder to reduce code
duplication.
- Update Python Runtimes to latest
- Add required 'jurisdictionStatus' field to license records, 
- No longer writing license 'status' field to DB, as it will be
calculated when read.

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #296