Feat/military affiliation records (#363)
We need to support licensees uploading documentation as evidence for
their military affiliation. We will be recording the military
affiliation upload history as records in the provider DynamoDB table
which will be used by the system to determine a user's current military
affiliation status. As part of this, we've added a new POST endpoint
which allows the user to specify a filename. When called, the API
generates a S3 pre-sign URL (active for 10 minutes) which the front end
will then use to upload that file into S3 directly. Once the file is
successfully uploaded, S3 will trigger a lambda function to complete the
initialization of the military affiliation record for that provider so
they will be eligible for any military discounts when purchasing
privileges.

When a user wants to end their military affiliation, the client will
call a PATCH endpoint that will set their military status to inactive.


### Requirements List
- No new requirements for deployment. All changes are backwards
compatible.

### Description List
- Add S3 bucket, which will be used to store licensee documents that
they upload into the system.
- Add event trigger to handle upload events to this bucket
- Add POST api endpoint for creating military affiliation records
- Add PATCH api endpoint for deactivating military affiliation status.
- Add Smoke tests which can be run locally against a sandbox environment
to verify the military document upload flow.
- Update Privilege purchase endpoint to check user's military
affiliation status when determining if they qualify for any military
discounts.
- Update provider data endpoint to return list of military affiliation
records as part of their profile information

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Local smoke tests
- Code review

Closes #[327](https://github.com/csg-org/CompactConnect/issues/327)