Feat/Record licensee id in privilege transactions (#368)
For reporting purposes, we must capture the user id of the licensee in
the transaction record in authorize.net. Our user ids are generated as
UUIDv4 strings, which are 36 characters long and contain dashes.
Originally, we had intended to use the customer id field authorize.net
provides for storing these. However, that field is limited to 20
alphanumeric characters (no dashes). The only field that will support
the length and dashes of our user ids is the order description field.
This field is readonly once the transaction is created, allows up to 255
characters, and is not restricted to alphanumeric characters.

This PR updates the Purchase client to pass in the licensee id into the
order description, so it can be referenced by our reporting which will
be implemented in a future ticket.

### Requirements List
- no new requirements. Using existing information.

### Description List
- Add required 'licensee_id' field when processing privilege purchases.

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #287