Feat/standardize timezone used to track datetimes (#369)
Previously, our system was using a non-standard timezone for recording
when records were created/updated. This could
potentially cause problems in the future should the needs of the system
change in regards to which timezone we need to
evaluate our records against. To correct this, this PR updates the
system to always use standard UTC timestamps when writing records to the
DB. This also converts the Date fields to DateTime fields, as this will
provide the granularity we will need to flexibly support changing
timezone requirements in the future.

The exception to this is license record fields, as these are defined by
the states that upload the records, and they do not include timestamps,
so we store whatever date value they give us.

We are releasing this change in a backwards compatible manner, in that
if a record has these fields currently recorded as date strings
'YYYY-MM-DD' we convert them to datetimes using a consistent timestamp.

### Requirements List
- No new requirements. These changes should be backwards compatible with
existing records.

### Description List
- using consistent UTC timezone to write records to the DB
- converting date fields to datetimes (except for license records)

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review

Closes #356