Replace current 'read' scope with 'readGeneral'
The previous 'read' compact scope is now replaced with the
'readGeneral' scope, which is now granted to every
staff user as part of the the pre-token generation. We also
add a '{compact}/{jurisdiction}.readGeneral' scope for all users
that have jurisdiction permissions. This will allow state admins
to query for all general information of licensees within their
jurisdiction.