Set API allowed origin headers (#461)
### Description List
For `Access-Control-Allowed-Origin` headers:
- Added specific/multiple origin support in OPTIONS preflight methods
- Added specific/multiple origin support for headers in all Lambda API
responses
- Added specific/ *single* origin support for headers in custom Gateway
Responses

Note that [*Gateway
Responses*](https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-gatewayResponse-definition.html)
are API responses that API Gateway itself returns, when it
short-circuits a request before delivering it to our Lambda back-end.
These responses are for requests that:
- Fail input validation for a POST body or query params,
- Are missing a valid authentication token, or
- Are missing required scopes for the endpoint.

These are specific 400, 401, and 403 responses in specific cases. In
those cases, API Gateway has limited support for us customizing the
content of the response. That customization does _not_ include
calculating a dynamic value based on the request `Origin` header, like
we are doing for the other endpoints/responses. Because of this
limitation, in the case of these Gateway Responses, we can only support
a single, static, value for the allowed origin header or, if we want to
support both a proper domain name and `http://localhost:3018` as origins
(like in the case of IA's test environment, where we want a common
deployed test UI and also for devs to be able to connect their local UI
to the shared test environment back-end), we have to still return a
`'*'` as allowed origin to support both at once. This limitation does
not constrain the production environment or any environments that will
have a single allowed origin, since we _can_ set the single origin in
the header as a static value.

### Testing List
- Code review
- Test browser behavior in IA's *test* environment

Closes #
Closes #165