Move DynamoDB writes into transaction
To improve robustness of system, this moves
the two writes we perform to the provider
records into a transaction write API call.
This will ensure that the whole operation
succeeds or is rolled back.

We split the cognito and dynamo calls into separate
try/except blocks to make the proper cleanup calls
as needed.