Feat/licensee registration (#460)
Provider need to be able to register in the system to create their
accounts and purchase privileges. This adds the
needed API endpoint and security infrastructure to support this.

The registration process is outlined as follows:
1. A state uploads a license record for a provider, which causes a
license record to be created
2. A provider registers through the registration page, specifying their
information as recorded on their license.
3. If the provided information matches a license record, and no one has
registered for an account under that license, a Cognito account is
created and an email invite is sent to them with login information.

### Requirements List
- 🚨 🚨 **BEFORE THIS CHANGE CAN BE MERGED!!!** 🚨 🚨CSG must configure
Google reCaptcha tokens for both the test and production environments.
The instructions for this setup are documented in the README changes as
part of this PR. **IF THIS IS MERGED BEFORE THAT SETUP, THE PIPELINE
WILL BE BROKEN!** Please ensure that this setup is completed before the
final approval on this PR.

- As part of this change, we have introduced a new GSI on the provider
table, which allows us to match a license record by a person's first and
last name. **This GSI will not match against any existing test records
in the test environment**, so new test records will need to be populated
in the provider table before this functionality can be tested in an
environment.

### Description List
- Added POST `/v1/provider-users/registration` **public facing**
endpoint to allow practitioners to register within the system.
- Capturing the home state selection that a practitioner selects when
registering with the system. This selection is now used on the backend
to determine which license is a practitioner's home state license.
- Updated privilege purchase endpoint to use this home state selection
record to pick best license.
- Update ingest pipeline to use home state selection to choose license
to associate with the provider
- Returning home state selection record in the provider detail response.

### Testing List
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Testing in local sandbox environment to verify functionality
- Code review

Closes #433

---------

Co-authored-by: Justin Frahm <justin@inspiringapps.com>