Feat/monthly reports (#508)
CSG has requested that we implement automated monthly transaction
reporting. This required several updates to increase the number of
transactions that could be handled within a single report. Previously we
were passing in the full csv report string to the email-service lambda
payload, which was prone to a 6MB file size limit, now we are storing
these reports in a compressed zip format in S3, and passing in the S3
key to the email service so it can pull down the file and attach the zip
file to the email directly. This adds the needed infrastructure to
support monthly reporting in a scaleable manner.

### Requirements List
- No new requirements, this change is backwards compatible with the
current system

### Description List
- Added event bridge rule to trigger transaction reporter lambda on the
first day of every month
- Added transaction reporting S3 bucket to store compiled csv report
files in a compressed format
- Update email-notification-service node lambda to pull down report
files from S3
- Fix formatting for reporting email body text to support markdown

### Testing List
- `yarn test:unit:all` should run without errors or warnings
- `yarn serve` should run without errors or warnings
- `yarn build` should run without errors or warnings
-  Python tests added for monthly reporting code verification
- Verification of monthly report generation for over 100k transactions
in sandbox environment.

Closes #493

---------

Co-authored-by: Justin Frahm <justin@inspiringapps.com>