Rename compactName field to compactAbbr (#558)
We are adding a new field for the full name of the Compact to the
configuration files, so we are refactoring the current field to denote
it is used for the compact abbreviation.

### Requirements List
- This change should be a non-breaking change, since we are updating all
code references to the field and updating
the database records.

### Description List
- renamed all original usage of `compactName` field to `compactAbbr`
- added full compact names to config files and schemas that reference
files.

### Testing List
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Code review
- verified privilege purchasing flow using sandbox environment

Closes #350