Feat/reporting updates (#542)
Updating the CSV reports based on feedback from compact EDs, the biggest
change being we are now including the associated privilege ids for each
transaction, so that admins can make the correlation between charges in
authorize.net and the compact connect system.

### Requirements List
- As part of this change, we are now updating the line item 'itemId' for
every privilege purchase with the prefix `priv:` to more readily
distinguish what is a privilege jurisdiction cost and what is compact
related fees. This means that older transactions which do not include
the `priv:` prefix will not show up in transaction reporting. Given that
the system is not live and this change is only related to reporting,
this will not impact any other functionality of the system.

### Description List
- Updated compact financial summary reports to include the following:
    - privilege purchase count by jurisdiction
    - total processed amount
    - change `Compact Fee` to `Administrative Fees`
- Updated transaction detail reports to include the following:
- Collected transaction fee if compact charged licensee transaction fees
   - Specify settlement date is UTC
- Associated privilege id for the transaction according to compact
connect
   - Transaction Status according to authorize.net

- Added compactTransactionId global secondary index for mapping
authorize.net transaction ids to compact connect privilege ids.

### Testing List
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- load testing script added for processing large numbers of transactions
- local testing in sandbox environment
- added unit/functional tests for reporting lambdas
- Code review

Closes #492