Feat/Support license type selection when purchasing privileges (#650)
As part of supporting multiple license types in the same home state,
this adds a `licenseType` field to the request body of the privilege
purchase endpoint. The endpoint now validates that the specified license
type matches one of the provider's home state licenses before proceeding
with the privilege purchase.

### Requirements List
- This is a backwards incompatible change, when this is merged the UI
will not be able to purchase privileges until [this PR for the
corresponding front end change is merged.
](https://github.com/csg-org/CompactConnect/pull/627)

### Description List
- Add required `licenseType` field to POST privilege purchase request
body
- Validation license type matches one of provider's home state licenses

### Testing List
- updated api schema snapshots
- updated smoke tests and verified change in sandbox env
- functional/unit test coverage added for validation logic
- Code review

Closes #572