Add field for enabling registration at compact/jurisdiction levels (#652)
Some compacts/jurisdictions will need time to onboard onto the system,
during which we do not want licensees being able
to register within that particular compact or jurisdiction. This adds a
new field `licenseeRegistrationEnabledForEnvironments` which is a list
of environments that allow registration within the respective
compact/jurisdiction.

### Requirements List
- No new requirements

### Description List
- Added `licenseeRegistrationEnabledForEnvironments` field to all
configuration files to gate registration
- Added checks in registration point to ensure registration is enabled
for selected compact/jurisdiction pair

### Testing List
- Functional tests added to 
- Code review

Closes #597