Tracking deactivation notes when priv is deactivated (#666)
When a privilege is deactivated, the staff user will provide a note as
to why the privilege is being deactivated. This adds support for storing
the note as well as information about the staff user in the deactivation
record.

In order to keep this change compatible with the frontend until it has
been updated, the 'deactivationNote' field has been set as optional.
TODO's have been added to be addressed once the frontend has been
updated to support it.

### Description List
- added 'deactivationDetails' field to privilege update record schema
- updated deactivation endpoint to get staff user information and store
it in the privilege update record
- Add migration script to add field to existing deactivation update
records

### Testing List
- For API configuration changes: CDK tests added/updated in
`backend/compact-connect/tests/unit/test_api.py`
- Functional tests updated
- Code review

Closes #566