Reference data event bus through SSM parameter (#731) (#736)
While testing the pipeline migration, we hit an issue where the event
bus needs to be renamed, as it is using the pipeline stack as part of
its name. This creates a cross stack reference issue since it is being
referenced by several other stacks, and you cannot change a stack export
value while it is still being referenced by other stacks. In order to
address this, this PR replaces the direct reference with an SSM
parameter, which can be changed without causing the stacks to fail
deployment. This PR will need to be merged and deployed first, before
the subsequent pipeline change can be run through.

This change is a precursor to the following tickets:
https://github.com/csg-org/CompactConnect/issues/705
https://github.com/csg-org/CompactConnect/issues/580
https://github.com/csg-org/CompactConnect/issues/424

Co-authored-by: Joshua Kravitz <jlkravitz@gmail.com>