Update compact config snapshot tests to track beta instead of test config
Now that we are actively adding states and making config changes to jurisdictions
The test environment snapshots for the compact config is causing more friction than
benefit. Beta and Prod are better candidates for snapshot tests, as they change less
frequently and we care more about the compacts/jurisdictions we are letting into
those environments. This replaces the test env snapshots with beta env snapshots
for tracking compact config.